experimentation with playwright
