+++ monitoring/package.json
@@ -18,9 +18,10 @@
     "dotenv": "^16.3.1",
     "inquirer": "^8.2.6",
     "playwright-aws-lambda": "^0.10.0",
-    "playwright-core": "^1.40.1",
     "tslib": "^2.6.2",
-    "uuid": "^9.0.1"
+    "uuid": "^9.0.1",
+		"@guardian/consent-management-platform": "13.8.0",
+		"playwright-core": "^1.40.1"
   },
   "devDependencies": {
     "@aws-sdk/client-lambda": "^3.490.0",
