+++ monitoring/src/check-page/common-functions.ts
@@ -6,7 +6,7 @@ import {
 	PutMetricDataCommand
 } from '@aws-sdk/client-cloudwatch';
 import { launchChromium } from 'playwright-aws-lambda';
-import type { Browser, BrowserContext, Page, Request } from 'playwright-core';
+import type { Browser, BrowserContext, Page } from 'playwright-core';
 import type { Config } from '../types';
 import { ELEMENT_ID } from '../types';
 
@@ -71,24 +71,6 @@ export const makeNewBrowser = async (debugMode: boolean): Promise<Browser> => {
 export const makeNewPage = async (context: BrowserContext): Promise<Page> => {
 	const page = await context.newPage();
 	return page;
-};
-
-/**
- * This function waits for the page to load
- * clicks the accept all button
- *
- * @param {Config} config
- * @param {Page} page
- * @param {string} textToPrintToConsole
- */
-export const clickAcceptAllCookies = async (config: Config, page: Page, textToPrintToConsole: string) => {
-
-	log_info(`Clicking on "${textToPrintToConsole}" on CMP`);
-
-	const acceptAllButton = page.frameLocator(ELEMENT_ID.CMP_CONTAINER).locator(ELEMENT_ID.TCFV2_FIRST_LAYER_ACCEPT_ALL);
-  	await acceptAllButton.click();
-
-	log_info(`Clicked on "${textToPrintToConsole}"`);
 };
 
 /**
@@ -163,56 +145,6 @@ export const clickRejectAllSecondLayer = async (config: Config, page: Page) => {
 	await page.frameLocator('[src*="' + config.iframeDomainSecondLayer + '"]').locator(ELEMENT_ID.TCFV2_SECOND_LAYER_REJECT_ALL).click();
 
 	log_info(`Clicking on reject all button: Complete`);
-};
-
-/**
- * This function checks for interaction with GAM
- * Using this after advice from Commercial to check that cookies were accepted as we otherwise do not interact with GAM
- * This has to be adjusted if anything in the interaction with GAM changes or we stop using GAM
- *
- * @param {Page} page
- * @return {*}  {Promise<void>}
- */
-export const checkTopAdHasLoaded = async (page: Page) => {
-	log_info(`Waiting for interaction with GAM: Start`);
-
-	const gamUrl = /https:\/\/securepubads.g.doubleclick.net\/gampad\/ads/;
-
-	const getEncodedParamsFromRequest = (
-		request: Request,
-		paramName: string,
-	): URLSearchParams | null => {
-		const url = new URL(request.url());
-		const param = url.searchParams.get(paramName);
-		if (!param) return null;
-		const paramDecoded = decodeURIComponent(param);
-		const searchParams = new URLSearchParams(paramDecoded);
-		return searchParams;
-	};
-
-	const assertOnSlotFromRequest = (request: Request, expectedSlot: string) => {
-		const isURL = request.url().match(gamUrl);
-		if (!isURL) return false;
-		const searchParams = getEncodedParamsFromRequest(request, 'prev_scp');
-		if (searchParams === null) return false;
-		const slot = searchParams.get('slot');
-		if (slot !== expectedSlot) return false;
-		return true;
-	};
-
-	const waitForGAMRequestForSlot = (page: Page, slotExpected: string) => {
-		return page.waitForRequest((request) =>
-			assertOnSlotFromRequest(request, slotExpected),
-		);
-	};
-
-	const gamRequestPromise = waitForGAMRequestForSlot(
-		page,
-		'top-above-nav',
-	);
-	await gamRequestPromise;
-
-	log_info(`Waiting for interaction with GAM: Complete`);
 };
 
 /**
